backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
python-multipart==0.0.6

# Include all core SRAG dependencies
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==3.0.1
torch==2.2.2+cpu
torchvision==0.17.2+cpu
chromadb==0.4.18
google-generativeai==0.3.0
